requirements.txt
openpyxl
pandas
oracledb
numpy
python-dotenv